+++ frontend/src/Pages/InitPage/InitPage.css
