+++ frontend/src/Pages/InitPage/InitPage.tsx
